+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - PR #4766 Update to use header-only NVTX v3 and remove need to link against nvtx.
 - PR #4716 Remove direct calls to RMM_ALLOC/RMM_FREE
 - PR #4765 Add in java support for sequence
+- PR #4660 Port `cudf::partition` api to python/cython
 
 ## Bug Fixes